@@ -5,28 +5,28 @@
     "license": "MIT",
     "type": "project",
     "require": {
-        "php": ">=7.2.5",
+        "php": "^7.3|^8.0",
         "ext-intl": "*",
         "ext-gd": "*",
-        "laravel/lumen-framework": "7.2.2",
-        "vlucas/phpdotenv": "4.1.8",
-        "illuminate/mail": "7.30.1",
-        "doctrine/dbal": "2.10.4",
+        "laravel/lumen-framework": "8.3.2",
+        "illuminate/contracts": "v8.75.0",
+        "illuminate/mail": "v8.57.0",
+        "doctrine/dbal": "^3.2",
         "fakturoid/fakturoid-php": "1.0.3",
-        "rikudou/czqrpayment": "4.0.2",
-        "endroid/qr-code": "3.9.6",
-        "loudislav/debatematch": "dev-master",
+        "rikudou/czqrpayment": "5.2.1",
+        "endroid/qr-code": "^4.4.1",
         "bugsnag/bugsnag-laravel": "2.21.0",
         "league/flysystem-aws-s3-v3": "^1.0"
     },
     "require-dev": {
-        "fzaninotto/faker": "1.9.2",
         "phpunit/phpunit": "8.5.13",
-        "mockery/mockery": "1.3.3"
+        "mockery/mockery": "1.3.3",
+        "fakerphp/faker": "^1.19"
     },
     "autoload": {
         "psr-4": {
-            "App\\": "app/"
+            "App\\": "app/",
+            "Database\\Factories\\": "database/factories/"
         }
     },
     "autoload-dev": {
